migrate render pass to render core
