+++ crates/thyllore-anim-core/src/animation_type.rs
@@ -1,0 +1,12 @@
+#[derive(Clone, Debug, PartialEq)]
+pub enum AnimationType {
+    None,
+    Skeletal,
+    Node,
+}
+
+impl Default for AnimationType {
+    fn default() -> Self {
+        Self::None
+    }
+}
