+++ crates/thyllore-anim-core/src/lib.rs
@@ -1,3 +1,4 @@
+mod animation_type;
 mod clip;
 mod constraint;
 pub mod editable;
@@ -5,6 +6,7 @@ pub mod keyframe_search;
 mod pose;
 pub mod spring_bone;
 
+pub use animation_type::AnimationType;
 pub use clip::*;
 pub use constraint::*;
 pub use pose::*;
