+++ crates/thyllore-vulkan-core/src/lib.rs
@@ -5,9 +5,11 @@ pub mod command;
 pub mod core;
 pub mod data;
 pub mod descriptor;
+pub mod frame_context;
 pub mod pipeline;
 pub mod raytracing;
 pub mod render;
+pub mod renderer;
 pub mod resource;
 pub mod vulkan;
 
@@ -15,8 +17,10 @@ pub use command::*;
 pub use core::*;
 pub use data::*;
 pub use descriptor::*;
+pub use frame_context::FrameRenderContext;
 pub use pipeline::*;
 pub use raytracing::*;
 pub use render::*;
+pub use renderer::*;
 pub use resource::*;
 pub use vulkan::*;
