+++ src/ecs/resource/depth_of_field.rs
@@ -1,16 +1,1 @@
-#[derive(Clone, Debug)]
-pub struct DepthOfField {
-    pub enabled: bool,
-    pub focus_distance: f32,
-    pub max_blur_radius: f32,
-}
-
-impl Default for DepthOfField {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            focus_distance: 10.0,
-            max_blur_radius: 8.0,
-        }
-    }
-}
+pub use thyllore_render_core::DepthOfField;
