+++ src/vulkanr/renderer/deferred/mod.rs
@@ -1,21 +1,13 @@
-mod auto_exposure;
-mod bloom;
 mod composite;
-mod dof;
 pub mod gbuffer;
 mod onion_skin;
 mod overlay_renderer;
 mod pass_recording;
 mod rayquery;
-mod tonemap;
 
-pub use auto_exposure::AutoExposurePass;
-pub use bloom::BloomPass;
 pub use composite::CompositePass;
-pub use dof::DofPass;
 pub use gbuffer::{create_gbuffer_framebuffer, GBufferPass};
 pub use onion_skin::OnionSkinRenderPass;
 pub use overlay_renderer::OverlayRenderer;
 pub use pass_recording::*;
 pub use rayquery::RayQueryPass;
-pub use tonemap::ToneMapPass;
